backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
openai>=1.25.0
anthropic>=0.25.0
google-generativeai>=0.8.0
